+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.4.10 (...)
+- support `FormPdoAuthentication` and make it the default
+- deprecate `FormAuthentcation`, new deploys will use `FormPdoAuthentication` 
+  by default
+
 ## 1.4.9 (2018-02-06)
 - set default for refresh token expiry to 180 days (instead of 6 months to sync
   with default of CA certificates)